chore(setup): move test classes
